Change RNA_raw_type_sizeof to return size_t
Avoids numeric overflow in some of the existing code.

bpy_rna.cc#foreach_parse_args has been updated to take a `size_t*`
argument since it was returning the result of #RNA_raw_type_sizeof. Its
use in bpy_rna.cc#foreach_getset has been updated to store the result in
a `size_t`.

+++ source/blender/makesrna/intern/rna_access.cc
@@ -4664,7 +4664,7 @@ int RNA_property_collection_raw_array(
   } \
   (void)0
 
-int RNA_raw_type_sizeof(RawPropertyType type)
+size_t RNA_raw_type_sizeof(RawPropertyType type)
 {
   switch (type) {
     case PROP_RAW_CHAR:
